<==== Mercurial data ====>
Branch: refactor_5059_edge_optimization

+++ common/src/core/dataprovider/cpull_media_stream_provider.cpp
@@ -136,7 +136,7 @@ void QnClientPullMediaStreamProvider::run()
         QnLiveStreamProvider* lp = dynamic_cast<QnLiveStreamProvider*>(this);
         if (videoData)
         {
-            m_stat[videoData->channelNumber].onData(videoData->dataSize());
+            m_stat[videoData->channelNumber].onData(static_cast<unsigned int>(videoData->dataSize()));
             if (lp)
                 lp->onGotVideoFrame(videoData);
         }
